@Scottmitch 2016-08-25
@normanmaurer @carl-mastrangelo - FYI. I think we can still accept up to 5 bytes even if we add in the result of the first byte. The approach does save us conditionals and unrolling the loop ensures we do only what is required.

Here are the byte code sizes for the two different versions of decodeULE128:
== manually unrolled
io.netty.handler.codec.http2.internal.hpack.Decoder::decodeULE128 (603 bytes)

== in loop form
io.netty.handler.codec.http2.internal.hpack.Decoder::decodeULE128 (140 bytes)

In a tight loop of 30000 iterations calling this method neither were inlined.

Our benchmarks don't report any major difference, but the change is not really significant in the grand scheme of things.

```
== new

Benchmark                     (limitToAscii)  (maxTableSize)  (sensitive)  (size)  Mode  Cnt       Score       Error  Units
HpackDecoderBenchmark.decode            true            4096         true   SMALL  avgt    5    2497.950 ±    95.966  ns/op
HpackDecoderBenchmark.decode            true            4096         true  MEDIUM  avgt    5   20010.398 ±   849.047  ns/op
HpackDecoderBenchmark.decode            true            4096         true   LARGE  avgt    5  374079.795 ± 34794.462  ns/op
HpackDecoderBenchmark.decode            true            4096        false   SMALL  avgt    5    2568.289 ±    65.099  ns/op
HpackDecoderBenchmark.decode            true            4096        false  MEDIUM  avgt    5   19696.710 ±   651.509  ns/op
HpackDecoderBenchmark.decode            true            4096        false   LARGE  avgt    5  366718.062 ± 13667.166  ns/op
HpackDecoderBenchmark.decode           false            4096         true   SMALL  avgt    5     411.391 ±    48.403  ns/op
HpackDecoderBenchmark.decode           false            4096         true  MEDIUM  avgt    5    1310.955 ±    46.909  ns/op
HpackDecoderBenchmark.decode           false            4096         true   LARGE  avgt    5   10139.187 ±   594.515  ns/op
HpackDecoderBenchmark.decode           false            4096        false   SMALL  avgt    5     473.132 ±    34.563  ns/op
HpackDecoderBenchmark.decode           false            4096        false  MEDIUM  avgt    5    1526.870 ±    99.320  ns/op
HpackDecoderBenchmark.decode           false            4096        false   LARGE  avgt    5   11289.348 ±   742.572  ns/op

== old

HpackDecoderBenchmark.decode            true            4096         true   SMALL  avgt    5    2478.468 ±   144.000  ns/op
HpackDecoderBenchmark.decode            true            4096         true  MEDIUM  avgt    5   20012.012 ±   607.477  ns/op
HpackDecoderBenchmark.decode            true            4096         true   LARGE  avgt    5  387318.233 ± 13614.705  ns/op
HpackDecoderBenchmark.decode            true            4096        false   SMALL  avgt    5    2527.243 ±    81.946  ns/op
HpackDecoderBenchmark.decode            true            4096        false  MEDIUM  avgt    5   19976.146 ±   873.261  ns/op
HpackDecoderBenchmark.decode            true            4096        false   LARGE  avgt    5  369174.480 ± 10646.162  ns/op
HpackDecoderBenchmark.decode           false            4096         true   SMALL  avgt    5     438.446 ±    16.675  ns/op
HpackDecoderBenchmark.decode           false            4096         true  MEDIUM  avgt    5    1412.593 ±    63.244  ns/op
HpackDecoderBenchmark.decode           false            4096         true   LARGE  avgt    5    9871.445 ±   429.030  ns/op
HpackDecoderBenchmark.decode           false            4096        false   SMALL  avgt    5     448.531 ±    33.905  ns/op
HpackDecoderBenchmark.decode           false            4096        false  MEDIUM  avgt    5    1507.985 ±   153.222  ns/op
HpackDecoderBenchmark.decode           false            4096        false   LARGE  avgt    5   11397.073 ±   687.688  ns/op
```