@Scottmitch 2016-09-02
@nmittler - comments resolved PTAL