@carl-mastrangelo 2016-09-02
@Scottmitch FYI (since github cant find the discussion), The C1 compiler says "callee is too large", the C2 one says "too big"  so you might not be looking at the most optimized inlining detail.