@nmittler 2016-09-02
@Scottmitch did you rerun the benchmarks?