@Scottmitch 2016-09-22
Here are the updated results ... the tests are a bit noisy on my local machine but I don't see any significant difference between either version. I will just go with the loop version since the byte code size is smaller and it is what we currently have.

```
this PR with loop

Benchmark                     (limitToAscii)  (maxTableSize)  (sensitive)  (size)  Mode  Cnt       Score       Error  Units
HpackDecoderBenchmark.decode            true            4096         true   SMALL  avgt    5    2512.026 ±   121.898  ns/op
HpackDecoderBenchmark.decode            true            4096         true  MEDIUM  avgt    5   20219.495 ±   302.784  ns/op
HpackDecoderBenchmark.decode            true            4096         true   LARGE  avgt    5  384906.788 ± 61173.300  ns/op
HpackDecoderBenchmark.decode            true            4096        false   SMALL  avgt    5    2541.537 ±   100.136  ns/op
HpackDecoderBenchmark.decode            true            4096        false  MEDIUM  avgt    5   20602.658 ±   239.226  ns/op
HpackDecoderBenchmark.decode            true            4096        false   LARGE  avgt    5  368972.633 ±  7979.227  ns/op
HpackDecoderBenchmark.decode           false            4096         true   SMALL  avgt    5     415.737 ±    21.008  ns/op
HpackDecoderBenchmark.decode           false            4096         true  MEDIUM  avgt    5    1321.411 ±    49.913  ns/op
HpackDecoderBenchmark.decode           false            4096         true   LARGE  avgt    5    9619.188 ±   559.867  ns/op
HpackDecoderBenchmark.decode           false            4096        false   SMALL  avgt    5     461.278 ±    32.666  ns/op
HpackDecoderBenchmark.decode           false            4096        false  MEDIUM  avgt    5    1480.829 ±    56.665  ns/op
HpackDecoderBenchmark.decode           false            4096        false   LARGE  avgt    5   10806.323 ±   615.366  ns/op

this PR loop unrolled

Benchmark                     (limitToAscii)  (maxTableSize)  (sensitive)  (size)  Mode  Cnt       Score       Error  Units
HpackDecoderBenchmark.decode            true            4096         true   SMALL  avgt    5    2434.397 ±   224.031  ns/op
HpackDecoderBenchmark.decode            true            4096         true  MEDIUM  avgt    5   19580.460 ±   549.175  ns/op
HpackDecoderBenchmark.decode            true            4096         true   LARGE  avgt    5  362169.235 ±  8518.092  ns/op
HpackDecoderBenchmark.decode            true            4096        false   SMALL  avgt    5    2446.413 ±    58.917  ns/op
HpackDecoderBenchmark.decode            true            4096        false  MEDIUM  avgt    5   19386.039 ±   485.982  ns/op
HpackDecoderBenchmark.decode            true            4096        false   LARGE  avgt    5  374434.923 ± 14721.887  ns/op
HpackDecoderBenchmark.decode           false            4096         true   SMALL  avgt    5     399.357 ±     8.426  ns/op
HpackDecoderBenchmark.decode           false            4096         true  MEDIUM  avgt    5    1318.147 ±    49.931  ns/op
HpackDecoderBenchmark.decode           false            4096         true   LARGE  avgt    5    9716.981 ±    87.016  ns/op
HpackDecoderBenchmark.decode           false            4096        false   SMALL  avgt    5     452.506 ±     8.102  ns/op
HpackDecoderBenchmark.decode           false            4096        false  MEDIUM  avgt    5    1483.872 ±    37.968  ns/op
HpackDecoderBenchmark.decode           false            4096        false   LARGE  avgt    5   10985.125 ±   287.254  ns/op

without this PR

Benchmark                     (limitToAscii)  (maxTableSize)  (sensitive)  (size)  Mode  Cnt       Score      Error  Units
HpackDecoderBenchmark.decode            true            4096         true   SMALL  avgt    5    2484.759 ±  113.955  ns/op
HpackDecoderBenchmark.decode            true            4096         true  MEDIUM  avgt    5   19647.768 ±  352.119  ns/op
HpackDecoderBenchmark.decode            true            4096         true   LARGE  avgt    5  360191.170 ± 8982.999  ns/op
HpackDecoderBenchmark.decode            true            4096        false   SMALL  avgt    5    2438.629 ±   43.502  ns/op
HpackDecoderBenchmark.decode            true            4096        false  MEDIUM  avgt    5   20311.295 ± 2069.466  ns/op
HpackDecoderBenchmark.decode            true            4096        false   LARGE  avgt    5  360407.129 ± 4317.484  ns/op
HpackDecoderBenchmark.decode           false            4096         true   SMALL  avgt    5     432.842 ±    7.385  ns/op
HpackDecoderBenchmark.decode           false            4096         true  MEDIUM  avgt    5    1376.419 ±   72.236  ns/op
HpackDecoderBenchmark.decode           false            4096         true   LARGE  avgt    5    9651.674 ±  234.388  ns/op
HpackDecoderBenchmark.decode           false            4096        false   SMALL  avgt    5     462.789 ±    7.912  ns/op
HpackDecoderBenchmark.decode           false            4096        false  MEDIUM  avgt    5    1477.598 ±   43.768  ns/op
HpackDecoderBenchmark.decode           false            4096        false   LARGE  avgt    5   10849.474 ±  253.683  ns/op

```